feat: 상세정보 컴포넌트 분리 (#36)
