+++ src/components/detail/Category.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Category = () => {
+  return;
+  <>
+    <Container></Container>
+  </>;
+};
+
+export default Category;
